requirements.txt
anyio==4.12.1
APScheduler==3.10.4
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.0
cot-reports==0.1.3
curl_cffi==0.13.0
cycler==0.12.1
exceptiongroup==1.3.1
fonttools==4.60.2
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.11
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
multitasking==0.0.12
numpy==2.0.2
packaging==26.0
pandas==2.3.3
peewee==3.19.0
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.33.4
pycparser==2.23
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==20.8
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
websockets==15.0.1
yfinance==0.2.66
zipp==3.23.0
